Refactored document upload to be application agnostic
By removing the store from aquifers and placing it into the common module, we can abstract the document upload process to not be tightly coupled to aquifers only. It now accepts a parameter that allows the caller to specify what type of file is being uploaded, allowing the code to be re-used.