Added ability to delete documents
Added a "Remove" link to all viewing document sections. If applicable, only when in edit mode of the record.

Added Django routes and views and augmented the document servcei to use the minio service to purge documents from the S3 store.

Does not yet refresh the file list when a document is removed.